+++ common/src/main/public/development.conf.json
@@ -2,7 +2,7 @@
   "environment": "DEVELOPMENT",
   "odbURI": "wss://lucuma-postgres-odb-dev.herokuapp.com/ws",
   "preferencesDBURI": "wss://user-prefs-development.herokuapp.com/v1/graphql",
-  "itcURI": "http://itc-development.herokuapp.com/itc",
+  "itcURI": "https://itc-development.herokuapp.com/itc",
   "sso": {
     "uri": "https://sso.gpp.lucuma.xyz",
     "readTimeoutSeconds": 10,
